Refactor onboarding_event fixture down to models_unit_test
